requirements.txt
# Framework principal
Flask==3.0.3

# Segurança e autenticação
Werkzeug==3.0.4

# Sessões, utilidades e templates
itsdangerous==2.2.0
Jinja2==3.1.4

# IA - Google Gemini
google-generativeai==0.8.3

# Manipulação de dados e cálculos
numpy==2.1.2
pandas==2.2.3
scikit-learn==1.5.2

# Requisições HTTP e variáveis de ambiente
requests==2.32.3
python-dotenv==1.0.1

# Extensões úteis do Flask (opcionais)
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1

google-generativeai==0.8.3
